+++ core/src/main/java/org/elasticsearch/index/query/GeoDistanceRangeQueryBuilder.java
@@ -24,6 +24,8 @@ import org.apache.lucene.spatial.geopoint.document.GeoPointField;
 import org.apache.lucene.spatial.geopoint.search.GeoPointDistanceRangeQuery;
 import org.apache.lucene.spatial.util.GeoDistanceUtils;
 import org.elasticsearch.Version;
+import org.elasticsearch.common.ParseField;
+import org.elasticsearch.common.ParsingException;
 import org.elasticsearch.common.Strings;
 import org.elasticsearch.common.geo.GeoDistance;
 import org.elasticsearch.common.geo.GeoPoint;
@@ -32,9 +34,11 @@ import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.unit.DistanceUnit;
 import org.elasticsearch.common.xcontent.XContentBuilder;
+import org.elasticsearch.common.xcontent.XContentParser;
 import org.elasticsearch.index.fielddata.IndexGeoPointFieldData;
 import org.elasticsearch.index.mapper.MappedFieldType;
 import org.elasticsearch.index.mapper.geo.BaseGeoPointFieldMapper;
+import org.elasticsearch.index.mapper.geo.GeoPointFieldMapper;
 import org.elasticsearch.index.mapper.geo.GeoPointFieldMapperLegacy;
 import org.elasticsearch.index.search.geo.GeoDistanceRangeQuery;
 
@@ -47,11 +51,32 @@ import static org.apache.lucene.spatial.util.GeoEncodingUtils.TOLERANCE;
 public class GeoDistanceRangeQueryBuilder extends AbstractQueryBuilder<GeoDistanceRangeQueryBuilder> {
 
     public static final String NAME = "geo_distance_range";
+    public static final ParseField QUERY_NAME_FIELD = new ParseField(NAME);
+
     public static final boolean DEFAULT_INCLUDE_LOWER = true;
     public static final boolean DEFAULT_INCLUDE_UPPER = true;
     public static final GeoDistance DEFAULT_GEO_DISTANCE = GeoDistance.DEFAULT;
     public static final DistanceUnit DEFAULT_UNIT = DistanceUnit.DEFAULT;
     public static final String DEFAULT_OPTIMIZE_BBOX = "memory";
+
+    public static final GeoDistanceRangeQueryBuilder PROTOTYPE = new GeoDistanceRangeQueryBuilder("_na_", new GeoPoint());
+
+    private static final ParseField FROM_FIELD = new ParseField("from");
+    private static final ParseField TO_FIELD = new ParseField("to");
+    private static final ParseField INCLUDE_LOWER_FIELD = new ParseField("include_lower");
+    private static final ParseField INCLUDE_UPPER_FIELD = new ParseField("include_upper");
+    private static final ParseField GT_FIELD = new ParseField("gt");
+    private static final ParseField GTE_FIELD = new ParseField("gte", "ge");
+    private static final ParseField LT_FIELD = new ParseField("lt");
+    private static final ParseField LTE_FIELD = new ParseField("lte", "le");
+    private static final ParseField UNIT_FIELD = new ParseField("unit");
+    private static final ParseField DISTANCE_TYPE_FIELD = new ParseField("distance_type");
+    private static final ParseField NAME_FIELD = new ParseField("_name");
+    private static final ParseField BOOST_FIELD = new ParseField("boost");
+    private static final ParseField OPTIMIZE_BBOX_FIELD = new ParseField("optimize_bbox");
+    private static final ParseField COERCE_FIELD = new ParseField("coerce", "normalize");
+    private static final ParseField IGNORE_MALFORMED_FIELD = new ParseField("ignore_malformed");
+    private static final ParseField VALIDATION_METHOD = new ParseField("validation_method");
 
     private final String fieldName;
 
@@ -69,8 +94,6 @@ public class GeoDistanceRangeQueryBuilder extends AbstractQueryBuilder<GeoDistan
     private String optimizeBbox = DEFAULT_OPTIMIZE_BBOX;
 
     private GeoValidationMethod validationMethod = GeoValidationMethod.DEFAULT;
-
-    static final GeoDistanceRangeQueryBuilder PROTOTYPE = new GeoDistanceRangeQueryBuilder("_na_", new GeoPoint());
 
     public GeoDistanceRangeQueryBuilder(String fieldName, GeoPoint point) {
         if (Strings.isEmpty(fieldName)) {
@@ -290,16 +313,227 @@ public class GeoDistanceRangeQueryBuilder extends AbstractQueryBuilder<GeoDistan
     protected void doXContent(XContentBuilder builder, Params params) throws IOException {
         builder.startObject(NAME);
         builder.startArray(fieldName).value(point.lon()).value(point.lat()).endArray();
-        builder.field(GeoDistanceRangeQueryParser.FROM_FIELD.getPreferredName(), from);
-        builder.field(GeoDistanceRangeQueryParser.TO_FIELD.getPreferredName(), to);
-        builder.field(GeoDistanceRangeQueryParser.INCLUDE_LOWER_FIELD.getPreferredName(), includeLower);
-        builder.field(GeoDistanceRangeQueryParser.INCLUDE_UPPER_FIELD.getPreferredName(), includeUpper);
-        builder.field(GeoDistanceRangeQueryParser.UNIT_FIELD.getPreferredName(), unit);
-        builder.field(GeoDistanceRangeQueryParser.DISTANCE_TYPE_FIELD.getPreferredName(), geoDistance.name().toLowerCase(Locale.ROOT));
-        builder.field(GeoDistanceRangeQueryParser.OPTIMIZE_BBOX_FIELD.getPreferredName(), optimizeBbox);
-        builder.field(GeoDistanceRangeQueryParser.VALIDATION_METHOD.getPreferredName(), validationMethod);
+        builder.field(FROM_FIELD.getPreferredName(), from);
+        builder.field(TO_FIELD.getPreferredName(), to);
+        builder.field(INCLUDE_LOWER_FIELD.getPreferredName(), includeLower);
+        builder.field(INCLUDE_UPPER_FIELD.getPreferredName(), includeUpper);
+        builder.field(UNIT_FIELD.getPreferredName(), unit);
+        builder.field(DISTANCE_TYPE_FIELD.getPreferredName(), geoDistance.name().toLowerCase(Locale.ROOT));
+        builder.field(OPTIMIZE_BBOX_FIELD.getPreferredName(), optimizeBbox);
+        builder.field(VALIDATION_METHOD.getPreferredName(), validationMethod);
         printBoostAndQueryName(builder);
         builder.endObject();
+    }
+
+    public static GeoDistanceRangeQueryBuilder fromXContent(QueryParseContext parseContext) throws IOException {
+        XContentParser parser = parseContext.parser();
+
+        XContentParser.Token token;
+
+        Float boost = null;
+        String queryName = null;
+        String currentFieldName = null;
+        GeoPoint point = null;
+        String fieldName = null;
+        Object vFrom = null;
+        Object vTo = null;
+        Boolean includeLower = null;
+        Boolean includeUpper = null;
+        DistanceUnit unit = null;
+        GeoDistance geoDistance = null;
+        String optimizeBbox = null;
+        boolean coerce = GeoValidationMethod.DEFAULT_LENIENT_PARSING;
+        boolean ignoreMalformed = GeoValidationMethod.DEFAULT_LENIENT_PARSING;
+        GeoValidationMethod validationMethod = null;
+
+        while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
+            if (token == XContentParser.Token.FIELD_NAME) {
+                currentFieldName = parser.currentName();
+            } else if (parseContext.isDeprecatedSetting(currentFieldName)) {
+                // skip
+            } else if (token == XContentParser.Token.START_ARRAY) {
+                if (fieldName == null) {
+                    if (point == null) {
+                        point = new GeoPoint();
+                    }
+                    GeoUtils.parseGeoPoint(parser, point);
+                    fieldName = currentFieldName;
+                } else {
+                    throw new ParsingException(parser.getTokenLocation(), "[" + GeoDistanceRangeQueryBuilder.NAME +
+                            "] field name already set to [" + fieldName + "] but found [" + currentFieldName + "]");
+                }
+            } else if (token == XContentParser.Token.START_OBJECT) {
+                // the json in the format of -> field : { lat : 30, lon : 12 }
+                if (fieldName == null) {
+                    fieldName = currentFieldName;
+                    if (point == null) {
+                        point = new GeoPoint();
+                    }
+                    GeoUtils.parseGeoPoint(parser, point);
+                } else {
+                    throw new ParsingException(parser.getTokenLocation(), "[" + GeoDistanceRangeQueryBuilder.NAME +
+                            "] field name already set to [" + fieldName + "] but found [" + currentFieldName + "]");
+                }
+            } else if (token.isValue()) {
+                if (parseContext.parseFieldMatcher().match(currentFieldName, FROM_FIELD)) {
+                    if (token == XContentParser.Token.VALUE_NULL) {
+                    } else if (token == XContentParser.Token.VALUE_STRING) {
+                        vFrom = parser.text(); // a String
+                    } else {
+                        vFrom = parser.numberValue(); // a Number
+                    }
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, TO_FIELD)) {
+                    if (token == XContentParser.Token.VALUE_NULL) {
+                    } else if (token == XContentParser.Token.VALUE_STRING) {
+                        vTo = parser.text(); // a String
+                    } else {
+                        vTo = parser.numberValue(); // a Number
+                    }
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, INCLUDE_LOWER_FIELD)) {
+                    includeLower = parser.booleanValue();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, INCLUDE_UPPER_FIELD)) {
+                    includeUpper = parser.booleanValue();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, GT_FIELD)) {
+                    if (token == XContentParser.Token.VALUE_NULL) {
+                    } else if (token == XContentParser.Token.VALUE_STRING) {
+                        vFrom = parser.text(); // a String
+                    } else {
+                        vFrom = parser.numberValue(); // a Number
+                    }
+                    includeLower = false;
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, GTE_FIELD)) {
+                    if (token == XContentParser.Token.VALUE_NULL) {
+                    } else if (token == XContentParser.Token.VALUE_STRING) {
+                        vFrom = parser.text(); // a String
+                    } else {
+                        vFrom = parser.numberValue(); // a Number
+                    }
+                    includeLower = true;
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, LT_FIELD)) {
+                    if (token == XContentParser.Token.VALUE_NULL) {
+                    } else if (token == XContentParser.Token.VALUE_STRING) {
+                        vTo = parser.text(); // a String
+                    } else {
+                        vTo = parser.numberValue(); // a Number
+                    }
+                    includeUpper = false;
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, LTE_FIELD)) {
+                    if (token == XContentParser.Token.VALUE_NULL) {
+                    } else if (token == XContentParser.Token.VALUE_STRING) {
+                        vTo = parser.text(); // a String
+                    } else {
+                        vTo = parser.numberValue(); // a Number
+                    }
+                    includeUpper = true;
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, UNIT_FIELD)) {
+                    unit = DistanceUnit.fromString(parser.text());
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, DISTANCE_TYPE_FIELD)) {
+                    geoDistance = GeoDistance.fromString(parser.text());
+                } else if (currentFieldName.endsWith(GeoPointFieldMapper.Names.LAT_SUFFIX)) {
+                    String maybeFieldName = currentFieldName.substring(0,
+                            currentFieldName.length() - GeoPointFieldMapper.Names.LAT_SUFFIX.length());
+                    if (fieldName == null || fieldName.equals(maybeFieldName)) {
+                        fieldName = maybeFieldName;
+                    } else {
+                        throw new ParsingException(parser.getTokenLocation(), "[" + GeoDistanceRangeQueryBuilder.NAME +
+                                "] field name already set to [" + fieldName + "] but found [" + currentFieldName + "]");
+                    }
+                    if (point == null) {
+                        point = new GeoPoint();
+                    }
+                    point.resetLat(parser.doubleValue());
+                } else if (currentFieldName.endsWith(GeoPointFieldMapper.Names.LON_SUFFIX)) {
+                    String maybeFieldName = currentFieldName.substring(0,
+                            currentFieldName.length() - GeoPointFieldMapper.Names.LON_SUFFIX.length());
+                    if (fieldName == null || fieldName.equals(maybeFieldName)) {
+                        fieldName = maybeFieldName;
+                    } else {
+                        throw new ParsingException(parser.getTokenLocation(), "[" + GeoDistanceRangeQueryBuilder.NAME +
+                                "] field name already set to [" + fieldName + "] but found [" + currentFieldName + "]");
+                    }
+                    if (point == null) {
+                        point = new GeoPoint();
+                    }
+                    point.resetLon(parser.doubleValue());
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, NAME_FIELD)) {
+                    queryName = parser.text();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, BOOST_FIELD)) {
+                    boost = parser.floatValue();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, OPTIMIZE_BBOX_FIELD)) {
+                    optimizeBbox = parser.textOrNull();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, COERCE_FIELD)) {
+                    coerce = parser.booleanValue();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, IGNORE_MALFORMED_FIELD)) {
+                    ignoreMalformed = parser.booleanValue();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, VALIDATION_METHOD)) {
+                    validationMethod = GeoValidationMethod.fromString(parser.text());
+                } else {
+                    if (fieldName == null) {
+                        if (point == null) {
+                            point = new GeoPoint();
+                        }
+                        point.resetFromString(parser.text());
+                        fieldName = currentFieldName;
+                    } else {
+                        throw new ParsingException(parser.getTokenLocation(), "[" + GeoDistanceRangeQueryBuilder.NAME +
+                                "] field name already set to [" + fieldName + "] but found [" + currentFieldName + "]");
+                    }
+                }
+            }
+        }
+
+        GeoDistanceRangeQueryBuilder queryBuilder = new GeoDistanceRangeQueryBuilder(fieldName, point);
+        if (boost != null) {
+            queryBuilder.boost(boost);
+        }
+
+        if (queryName != null) {
+            queryBuilder.queryName(queryName);
+        }
+
+        if (vFrom != null) {
+            if (vFrom instanceof Number) {
+                queryBuilder.from((Number) vFrom);
+            } else {
+                queryBuilder.from((String) vFrom);
+            }
+        }
+
+        if (vTo != null) {
+            if (vTo instanceof Number) {
+                queryBuilder.to((Number) vTo);
+            } else {
+                queryBuilder.to((String) vTo);
+            }
+        }
+
+        if (includeUpper != null) {
+            queryBuilder.includeUpper(includeUpper);
+        }
+
+        if (includeLower != null) {
+            queryBuilder.includeLower(includeLower);
+        }
+
+        if (unit != null) {
+            queryBuilder.unit(unit);
+        }
+
+        if (geoDistance != null) {
+            queryBuilder.geoDistance(geoDistance);
+        }
+
+        if (optimizeBbox != null) {
+            queryBuilder.optimizeBbox(optimizeBbox);
+        }
+
+        if (validationMethod != null) {
+            // if validation method is set explicitly ignore deprecated coerce/ignore malformed fields if any
+            queryBuilder.setValidationMethod(validationMethod);
+        } else {
+            queryBuilder.setValidationMethod(GeoValidationMethod.infer(coerce, ignoreMalformed));
+        }
+        return queryBuilder;
     }
 
     @Override
