+++ core/src/main/java/org/elasticsearch/index/query/TypeQueryBuilder.java
@@ -19,28 +19,30 @@
 
 package org.elasticsearch.index.query;
 
-import org.apache.lucene.index.Term;
 import org.apache.lucene.search.MatchNoDocsQuery;
 import org.apache.lucene.search.Query;
-import org.apache.lucene.search.TermQuery;
 import org.apache.lucene.util.BytesRef;
+import org.elasticsearch.common.ParseField;
+import org.elasticsearch.common.ParsingException;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.lucene.BytesRefs;
 import org.elasticsearch.common.xcontent.XContentBuilder;
+import org.elasticsearch.common.xcontent.XContentParser;
 import org.elasticsearch.index.mapper.DocumentMapper;
-import org.elasticsearch.index.mapper.internal.TypeFieldMapper;
 
 import java.io.IOException;
 import java.util.Objects;
 
 public class TypeQueryBuilder extends AbstractQueryBuilder<TypeQueryBuilder> {
-
     public static final String NAME = "type";
+    public static final ParseField QUERY_NAME_FIELD = new ParseField(NAME);
+
+    private static final ParseField VALUE_FIELD = new ParseField("value");
 
     private final BytesRef type;
 
-    static final TypeQueryBuilder PROTOTYPE = new TypeQueryBuilder("type");
+    public static final TypeQueryBuilder PROTOTYPE = new TypeQueryBuilder("type");
 
     public TypeQueryBuilder(String type) {
         if (type == null) {
@@ -63,10 +65,49 @@ public class TypeQueryBuilder extends AbstractQueryBuilder<TypeQueryBuilder> {
     @Override
     protected void doXContent(XContentBuilder builder, Params params) throws IOException {
         builder.startObject(NAME);
-        builder.field(TypeQueryParser.VALUE_FIELD.getPreferredName(), type.utf8ToString());
+        builder.field(VALUE_FIELD.getPreferredName(), type.utf8ToString());
         printBoostAndQueryName(builder);
         builder.endObject();
     }
+
+    public static TypeQueryBuilder fromXContent(QueryParseContext parseContext) throws IOException {
+        XContentParser parser = parseContext.parser();
+        BytesRef type = null;
+
+        String queryName = null;
+        float boost = AbstractQueryBuilder.DEFAULT_BOOST;
+
+        String currentFieldName = null;
+        XContentParser.Token token;
+        while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
+            if (token == XContentParser.Token.FIELD_NAME) {
+                currentFieldName = parser.currentName();
+            } else if (token.isValue()) {
+                if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.NAME_FIELD)) {
+                    queryName = parser.text();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.BOOST_FIELD)) {
+                    boost = parser.floatValue();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, VALUE_FIELD)) {
+                    type = parser.utf8Bytes();
+                } else {
+                    throw new ParsingException(parser.getTokenLocation(),
+                            "[" + TypeQueryBuilder.NAME + "] filter doesn't support [" + currentFieldName + "]");
+                }
+            } else {
+                throw new ParsingException(parser.getTokenLocation(),
+                        "[" + TypeQueryBuilder.NAME + "] filter doesn't support [" + currentFieldName + "]");
+            }
+        }
+
+        if (type == null) {
+            throw new ParsingException(parser.getTokenLocation(),
+                    "[" + TypeQueryBuilder.NAME + "] filter needs to be provided with a value for the type");
+        }
+        return new TypeQueryBuilder(type)
+                .boost(boost)
+                .queryName(queryName);
+    }
+
 
     @Override
     public String getWriteableName() {
